# docs/swagger.yaml
basePath: /
definitions:
  controllers.GameRequest:
    properties:
      code:
        example: ML
        type: string
      name:
        example: Mobile Legends
        type: string
    type: object
  controllers.LoginRequest:
    properties:
      email:
        example: budi@example.com
        type: string
      password:
        example: password123
        type: string
    type: object
...
  controllers.RegisterRequest:
    properties:
      email:
        example: budi@example.com
        type: string
      name:
        example: Budi
        type: string
      password:
        example: password123
        type: string
    type: object
  utils.APIResponse:
    properties:
      data:
        description: Data yang dikirim API.
      message:
        description: Message dari API.
        type: string
      status:
        description: |-
          Status response.
          Contoh: success / error
        type: string
    type: object
host: localhost:9000
info:
  contact: {}
  description: REST API untuk mengelola game, product, provider, dan order top-up.
  title: Topup Games API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Menghasilkan access token untuk request terproteksi.
      parameters:
      - description: Data login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Login user
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Membuat akun customer baru.
      parameters:
      - description: Data registrasi
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mendaftarkan user
      tags:
      - auth
  /api/games:
    get:
      description: Mengambil daftar seluruh game.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Mengambil semua game
      tags:
      - games
    post:
      consumes:
      - application/json
      description: Membuat game baru.
      parameters:
      - description: Data game
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.GameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Membuat game
      tags:
      - games
  /api/games/{id}:
    delete:
      description: Menghapus game secara soft delete berdasarkan ID.
      parameters:
      - description: ID game
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Menghapus game
      tags:
      - games
    get:
      description: Mengambil satu game berdasarkan ID.
      parameters:
      - description: ID game
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Mengambil game berdasarkan ID
      tags:
      - games
    put:
      consumes:
      - application/json
      description: Mengubah data game berdasarkan ID.
      parameters:
      - description: ID game
        in: path
        name: id
        required: true
        type: integer
      - description: Data game
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.GameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Mengubah game
      tags:
      - games
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Masukkan token dengan format "Bearer {token}".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"